kapsule-studio-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.0
google-cloud-aiplatform==1.38.0
google-cloud-secret-manager==2.16.4
python-multipart==0.0.6
python-dotenv==1.0.0
ffmpeg-python==0.2.0
pydantic==2.5.0
requests==2.31.0
google-auth==2.23.4

